refactor: 프로젝트 명 변경
