+++ be/withu/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'boomerang'
+rootProject.name = 'withu'
