--- be/withu/src/main/java/boomerang/global/response/ErrorCode.java
@@ -1,43 +1,0 @@
-package boomerang.global.response;
-
-import org.springframework.http.HttpStatus;
-
-public enum ErrorCode {
-    // Global
-    UNEXPECTED_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "EG001", "Unexpected Error"),
-    ACCESS_TOKEN_NOT_EXISTS_ERROR(HttpStatus.BAD_REQUEST, "EG002", "Access Token Not Exists Error"),
-    JWT_ERROR(HttpStatus.UNAUTHORIZED, "EG003", "JWT token is not valid"),
-
-    // Template
-    TEMPLATE_NOT_FOUND_ERROR(HttpStatus.BAD_REQUEST, "EM001", "Member Not Found Error"),
-    DUPLICATE_EMAIL_ERROR(HttpStatus.BAD_REQUEST, "EM002", "Duplicate Email Error"),
-    DUPLICATE_NICKNAME_ERROR(HttpStatus.BAD_REQUEST, "EM003", "Duplicate Nickname Error"),
-
-    ;
-    private final HttpStatus status;
-    private final String code;
-    private final String message;
-
-    ErrorCode(HttpStatus status, String code, String message) {
-        this.status = status;
-        this.code = code;
-        this.message = message;
-    }
-
-    public HttpStatus getStatus() {
-        return status;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    @Override
-    public String toString() {
-        return code + " : " + message;
-    }
-}
