--- be/withu/src/main/java/boomerang/global/response/ErrorResponseDto.java
@@ -1,7 +1,0 @@
-package boomerang.global.response;
-
-public record ErrorResponseDto(String code, String message) {
-    public ErrorResponseDto(ErrorCode errorCode) {
-        this(errorCode.getCode(), errorCode.getMessage());
-    }
-}
