--- be/withu/src/main/java/boomerang/global/response/SimpleResultResponseDto.java
@@ -1,7 +1,0 @@
-package boomerang.global.response;
-
-public record SimpleResultResponseDto(String code, String message) {
-    public SimpleResultResponseDto(ResultCode resultCode) {
-        this(resultCode.getCode(), resultCode.getMessage());
-    }
-}
