--- be/withu/src/main/java/boomerang/template/exception/TemplateValidException.java
@@ -1,11 +1,0 @@
-package boomerang.template.exception;
-
-import boomerang.global.exception.BusinessException;
-import boomerang.global.exception.DomainValidationException;
-import boomerang.global.response.ErrorCode;
-
-public class TemplateValidException extends DomainValidationException {
-    public TemplateValidException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
