--- be/withu/src/main/java/boomerang/template/repository/TemplateRepository.java
@@ -1,8 +1,0 @@
-package boomerang.template.repository;
-
-import boomerang.global.MyCrudRepository;
-import boomerang.template.domain.TemplateDomain;
-
-public interface TemplateRepository extends MyCrudRepository<TemplateDomain, Long> {
-    boolean existsById(Long id);
-}
