--- be/withu/src/main/java/boomerang/template/service/TemplateService.java
@@ -1,47 +1,0 @@
-package boomerang.template.service;
-
-import boomerang.template.domain.TemplateDomain;
-import boomerang.template.dto.TemplateServiceDto;
-import boomerang.template.exception.TemplateNotFoundException;
-import boomerang.template.repository.TemplateRepository;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class TemplateService {
-    private final TemplateRepository templateRepository;
-
-    public TemplateService(TemplateRepository templateRepository) {
-        this.templateRepository = templateRepository;
-    }
-
-    public List<TemplateDomain> getAllTemplateDomains() {
-        return templateRepository.findAll();
-    }
-
-    public TemplateDomain getTemplateDomainById(Long id) {
-        return templateRepository.findById(id)
-                .orElseThrow(TemplateNotFoundException::new);
-    }
-
-    public TemplateDomain createTemplateDomain(TemplateServiceDto templateCreateServiceDto) {
-        return templateRepository.save(templateCreateServiceDto.toTemplateDomain());
-    }
-
-    public TemplateDomain updateTemplateDomain(TemplateServiceDto templateCreateServiceDto) {
-        validateTemplateDomainExists(templateCreateServiceDto.getId());
-        return templateRepository.save(templateCreateServiceDto.toTemplateDomain());
-    }
-
-    public void deleteTemplateDomain(Long id) {
-        validateTemplateDomainExists(id);
-        templateRepository.deleteById(id);
-    }
-
-    private void validateTemplateDomainExists(Long id) {
-        if (!templateRepository.existsById(id)) {
-            throw new TemplateNotFoundException();
-        }
-    }
-}
