+++ be/withu/src/main/java/withu/global/exception/AccessTokenNotExistsException.java
@@ -1,0 +1,9 @@
+package withu.global.exception;
+
+import withu.global.response.ErrorCode;
+
+public class AccessTokenNotExistsException extends BusinessException {
+    public AccessTokenNotExistsException() {
+        super(ErrorCode.ACCESS_TOKEN_NOT_EXISTS_ERROR);
+    }
+}
