+++ be/withu/src/main/java/withu/global/response/ErrorResponseDto.java
@@ -1,0 +1,7 @@
+package withu.global.response;
+
+public record ErrorResponseDto(String code, String message) {
+    public ErrorResponseDto(ErrorCode errorCode) {
+        this(errorCode.getCode(), errorCode.getMessage());
+    }
+}
