+++ be/withu/src/main/java/withu/global/response/SimpleResultResponseDto.java
@@ -1,0 +1,7 @@
+package withu.global.response;
+
+public record SimpleResultResponseDto(String code, String message) {
+    public SimpleResultResponseDto(ResultCode resultCode) {
+        this(resultCode.getCode(), resultCode.getMessage());
+    }
+}
