+++ be/withu/src/main/java/withu/template/domain/TemplateDomain.java
@@ -1,0 +1,47 @@
+package withu.template.domain;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+import org.hibernate.Hibernate;
+
+import java.util.Objects;
+
+@Entity
+@Table(name = "template")
+@Getter
+@Setter
+public class TemplateDomain {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Embedded
+    private TemplateColumn1 templateColumn1;
+
+    @Embedded
+    private TemplateColumn2 templateColumn2;
+
+    public TemplateDomain() {
+    }
+
+    public TemplateDomain(Long id, TemplateColumn1 templateColumn1, TemplateColumn2 templateColumn2) {
+        this.id = id;
+        this.templateColumn1 = templateColumn1;
+        this.templateColumn2 = templateColumn2;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o))
+            return false;
+        TemplateDomain item = (TemplateDomain) o;
+        return Objects.equals(id, item.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id);
+    }
+}
