+++ be/withu/src/main/java/withu/template/dto/TemplateCreateResponseDto.java
@@ -1,0 +1,17 @@
+package withu.template.dto;
+
+import withu.template.domain.TemplateColumn1;
+import withu.template.domain.TemplateColumn2;
+import lombok.Getter;
+
+@Getter
+public class TemplateCreateResponseDto {
+    private TemplateColumn1 templateColumn1;
+    private TemplateColumn2 templateColumn2;
+
+    // 생성자
+    public TemplateCreateResponseDto(TemplateColumn1 templateColumn1, TemplateColumn2 templateColumn2) {
+        this.templateColumn1 = templateColumn1;
+        this.templateColumn2 = templateColumn2;
+    }
+}
