+++ be/withu/src/main/java/withu/template/repository/TemplateRepository.java
@@ -1,0 +1,8 @@
+package withu.template.repository;
+
+import withu.global.MyCrudRepository;
+import withu.template.domain.TemplateDomain;
+
+public interface TemplateRepository extends MyCrudRepository<TemplateDomain, Long> {
+    boolean existsById(Long id);
+}
